Code clean, fixes and better login
Workaround moved to their own  clase and code file.
Added logging.
Fix for condition evaluation.
